requirements.txt
fastapi
uvicorn
pandas
prophet
sentence-transformers
faiss-cpu
transformers